Networking: [Layer 4] TCP + [Layer 7] HTTP
Author:    Cytroyd <cytroyd@gmail.com>
Committer: Cytroyd <cytroyd@gmail.com>

+++ include/drivers/amd_am79c973.h
@@ -104,7 +104,7 @@ namespace rexos {
                 rexos::common::uint64_t GetMACAddress();
                 // Instead of using DHCP, we set the IP address manually
                 void SetIPAddress(common::uint32_t IP_BE);
-                common::uint64_t GetIPAddress();
+                common::uint32_t GetIPAddress();
         };
     }
